requirements.txt
streamlit>=1.56.0
fastapi>=0.116.1
uvicorn>=0.35.0
supabase>=2.28.3
openai>=2.30.0
pydantic>=2.12.5
email-validator>=2.2.0
plotly>=6.6.0
pandas>=3.0.2
httpx>=0.28.1
python-dotenv>=1.2.2
Faker>=40.12.0
streamlit-option-menu>=0.4.0
